- Added batch script linear search - Re-structure each code to separated folder
