+++ c-linear-search/README.md
@@ -1,0 +1,2 @@
+# c-linear-search
+*Example Linear Search in C*
