+++ cpp-linear-search/README.md
@@ -1,0 +1,2 @@
+# cpp-linear-search
+*Example Linear Search in Cpp*
