+++ golang-linear-search/README.md
@@ -1,0 +1,2 @@
+# golang-linear-search
+*Example Linear Search in Golang*
