+++ golang-linear-search/golang-linear-search.go
