+++ haskell-linear-search/README.md
@@ -1,0 +1,2 @@
+# haskell-linear-search
+*Example Linear Search in Haskell*
