+++ haskell-linear-search/haskell-linear-search.hs
