+++ java-linear-search/README.md
@@ -1,0 +1,2 @@
+# java-linear-search
+*Example Linear Search in Java*
