+++ javascript-linear-search/README.md
@@ -1,0 +1,2 @@
+# javascript-linear-search
+*Example Linear Search in Javascript*
