+++ javascript-linear-search/javascript-linear-search.js
