+++ kotlin-linear-search/README.md
@@ -1,0 +1,2 @@
+# kotlin-linear-search
+*Example Linear Search in Kotlin*
