+++ kotlin-linear-search/kotlin-linear-search.kt
