+++ lisp-linear-search/README.md
@@ -1,0 +1,2 @@
+# lisp-linear-search
+*Example Linear Search in Lisp*
