+++ lisp-linear-search/lisp-linear-search.lisp
