+++ php-linear-search/README.md
@@ -1,0 +1,2 @@
+# php-linear-search
+*Example Linear Search in Php*
