+++ ruby-linear-search/README.md
@@ -1,0 +1,2 @@
+# ruby-linear-search
+*Example Linear Search in Ruby*
