+++ ruby-linear-search/ruby-linear-search-2.rb
