+++ ruby-linear-search/ruby-linear-search.rb
