+++ rust-linear-search/README.md
@@ -1,0 +1,2 @@
+# rust-linear-search
+*Example Linear Search in Rust*
