+++ rust-linear-search/rust-linear-search.rs
