+++ scala-linear-search/README.md
@@ -1,0 +1,2 @@
+# scala-linear-search
+*Example Linear Search in Scala*
