+++ scala-linear-search/scala-linear-search.scala
