+++ swift-linear-search/README.md
@@ -1,0 +1,2 @@
+# swift-linear-search
+*Example Linear Search in Swift*
